Circuit PSI conversion
Convert datatypes in circuit psi code to uint8 for use of smaller fields.
Begin integration with main code.

+++ extern/MPCHonestMajority/CircuitPSI.h
@@ -65,13 +65,13 @@ class CircuitPSI : public ProtocolParty<FieldType>{
 		string myInputFile, myOutputFile;
 
 		CircuitPSI(int argc, char* argv[]);
-		CircuitPSI(int argc, char* argv[], vector<uint64_t>& bins, uint64_t nbins);
+		CircuitPSI(int argc, char* argv[], vector<uint8_t>& bins, uint64_t nbins);
 
 		//read num_bins MPSI inputs
 		void readMPSIInputs();
-		void readMPSIInputs(vector<vector<uint64_t>>& bins, uint64_t nbins);
+		void readMPSIInputs(vector<vector<uint8_t>>& bins, uint64_t nbins);
 
-		void convertSharestoFieldType(vector<uint64_t>& bins, vector<FieldType>& shares, uint64_t nbins);
+		void convertSharestoFieldType(vector<uint8_t>& bins, vector<FieldType>& shares, uint64_t nbins);
 
 		//perform MPSI
 		void runMPSI();
@@ -156,7 +156,7 @@ template <class FieldType> CircuitPSI<FieldType>::CircuitPSI(int argc, char* arg
 				triple_ctr = prime_bitlen + sindex.size() - 2;
 }
 
-template <class FieldType> CircuitPSI<FieldType>::CircuitPSI(int argc, char* argv[], vector<uint64_t>& bins, uint64_t nbins) : ProtocolParty<FieldType>(argc, argv) {
+template <class FieldType> CircuitPSI<FieldType>::CircuitPSI(int argc, char* argv[], vector<uint8_t>& bins, uint64_t nbins) : ProtocolParty<FieldType>(argc, argv) {
         //The call to ProtocolParty constructor initializes inherited variables
         // N, T, m_partyID, as well as the VDM matrix and related vectors.
 
@@ -193,7 +193,7 @@ template <class FieldType> CircuitPSI<FieldType>::CircuitPSI(int argc, char* arg
 template <class FieldType> void CircuitPSI<FieldType>::readMPSIInputs() {
         ifstream myfile;
         //long long int input;
-	uint64_t input;
+	uint8_t input;
         uint64_t i = 0;
         myfile.open(myInputFile);
         do {
@@ -220,8 +220,8 @@ template <class FieldType> void CircuitPSI<FieldType>::readMPSIInputs() {
 }
 
 //read num_bins MPSI inputs
-template <class FieldType> void CircuitPSI<FieldType>::readMPSIInputs(vector<vector<uint64_t>>& bins, uint64_t nbins) {
-  	  uint64_t input;
+template <class FieldType> void CircuitPSI<FieldType>::readMPSIInputs(vector<vector<uint8_t>>& bins, uint64_t nbins) {
+  	  uint8_t input;
     	uint64_t i = 0;
 	    uint64_t j = 0;
 
@@ -253,8 +253,8 @@ template <class FieldType> void CircuitPSI<FieldType>::readMPSIInputs(vector<vec
 }
 
 //convert shares to field type
-template <class FieldType> void CircuitPSI<FieldType>::convertSharestoFieldType(vector<uint64_t>& bins, vector<FieldType>& shares, uint64_t nbins) {
-  	uint64_t input;
+template <class FieldType> void CircuitPSI<FieldType>::convertSharestoFieldType(vector<uint8_t>& bins, vector<FieldType>& shares, uint64_t nbins) {
+  	uint8_t input;
 	uint64_t j = 0;
 	for(int i=0; i<nbins; i++) {
 		input = bins[j++];
